refactor drivers to follow naming conventions
