+++ tests/Stubs/SecondDriverStub/Driver.php
@@ -1,7 +1,9 @@
 <?php
-namespace Michaels\Spider\Test\Stubs;
+namespace Michaels\Spider\Test\Stubs\SecondDriverStub;
 
-class SecondDriverStub extends DriverStub
+use Michaels\Spider\Test\Stubs\FirstDriverStub\Driver as FirstDriver;
+
+class Driver extends FirstDriver
 {
     public function open(array $credentials, array $config = [])
     {
